lab6-eleição-coordenação-distribuída/pyproject.toml
[tool.poetry]
name = "lab6-eleição-coordenada-distrubuída"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "lab6_eleição_coordenada_distrubuída"}]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
